requirements.txt
numpy==2.3.2
Pillow==11.3.0
streamlit
tensorflow==2.18.0

